GPS location fetch moved to ExplorerActivity.
