+++ app/src/main/java/android/nik/virtualgeocaching/activities/ExplorerActivity.java
@@ -1,22 +1,34 @@
 package android.nik.virtualgeocaching.activities;
 
+import android.Manifest;
 import android.content.Intent;
+import android.content.pm.PackageManager;
+import android.location.Criteria;
+import android.location.Location;
+import android.location.LocationListener;
+import android.location.LocationManager;
 import android.nik.virtualgeocaching.model.Chest;
 import android.nik.virtualgeocaching.model.Map;
 import android.nik.virtualgeocaching.adapters.MapAdapter;
 import android.nik.virtualgeocaching.R;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
+import android.support.v4.app.ActivityCompat;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ListView;
+import android.widget.TextView;
+import android.widget.Toast;
 
+import com.google.android.gms.maps.model.LatLng;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
+
+import org.w3c.dom.Text;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -25,12 +37,17 @@ import java.util.List;
  * Created by Laszlo on 3/24/2017.
  */
 
-public class ExplorerActivity extends AppCompatActivity implements View.OnClickListener {
+public class ExplorerActivity extends AppCompatActivity implements View.OnClickListener, LocationListener {
 
     Map map;
     ListView explorerChestList;
     MapAdapter mapAdapter;
     private DatabaseReference mDatabase;
+    LocationManager locationManager;
+    String provider;
+    double longitude;
+    double latitude;
+    private TextView positionTextView;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
@@ -59,10 +76,34 @@ public class ExplorerActivity extends AppCompatActivity implements View.OnClickL
 
             }
         });
+        //VIEWS
+        positionTextView = (TextView) findViewById(R.id.positionText);
         //BUTTONS
         findViewById(R.id.adder_button).setOnClickListener(this);
 
+        //GPS
+        if (ActivityCompat.checkSelfPermission(this, Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED && ActivityCompat.checkSelfPermission(this, Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
+            ActivityCompat.requestPermissions(this,new String[]{Manifest.permission.ACCESS_COARSE_LOCATION,Manifest.permission.ACCESS_FINE_LOCATION},200);
+        }
+        locationManager = (LocationManager) getSystemService(LOCATION_SERVICE);
+        Criteria criteria = new Criteria();
+        provider = locationManager.getBestProvider(criteria, false);
 
+        if (provider != null && !provider.equalsIgnoreCase("")) {
+
+            Location location = locationManager.getLastKnownLocation(provider);
+            locationManager.requestLocationUpdates(LocationManager.NETWORK_PROVIDER, 15000,1,this);
+
+            if(location != null)
+                onLocationChanged(location);
+            else
+                Toast.makeText(ExplorerActivity.this, "No location provider found.", Toast.LENGTH_SHORT).show();
+        }
+
+        //GPS assign
+        positionTextView.setText(String.format("lat: %s long: %s",String.valueOf(latitude),String.valueOf(longitude)));
+
+        //CHESTLIST
         final ListView chestList = (ListView)findViewById(R.id.container_list);
         this.explorerChestList = chestList;
         mapAdapter = new MapAdapter(map);
@@ -94,6 +135,8 @@ public class ExplorerActivity extends AppCompatActivity implements View.OnClickL
         if(i == R.id.adder_button){
             Intent createChestActivity = new Intent(ExplorerActivity.this, CreateChestActivity.class);
             createChestActivity.putExtra("map",map);
+            createChestActivity.putExtra("explorerLatitude",latitude);
+            createChestActivity.putExtra("explorerLongitude",longitude);
             //startActivity(createChestActivity);
             startActivityForResult(createChestActivity, 100);
         }
@@ -114,5 +157,30 @@ public class ExplorerActivity extends AppCompatActivity implements View.OnClickL
     private void reloadChestList() {
         mapAdapter.setMap(map);
         mapAdapter.notifyDataSetChanged();
+    }
+    private LatLng getLocation()
+    {
+        return new LatLng(latitude,longitude);
+    }
+
+    @Override
+    public void onLocationChanged(Location location) {
+        longitude = location.getLongitude();
+        latitude = location.getLatitude();
+    }
+
+    @Override
+    public void onStatusChanged(String provider, int status, Bundle extras) {
+
+    }
+
+    @Override
+    public void onProviderEnabled(String provider) {
+
+    }
+
+    @Override
+    public void onProviderDisabled(String provider) {
+
     }
 }
