+++ app/src/main/res/layout/activity_explorer.xml
@@ -6,7 +6,12 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     tools:context="android.nik.virtualgeocaching.activities.ExplorerActivity">
-
+    <android.support.v7.widget.AppCompatTextView
+        android:id="@+id/positionText"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_weight="1"
+        />
     <ListView
         android:id="@+id/container_list"
         android:layout_width="match_parent"
